Bug fixes to cater for real time movement
Bug fixes to stop game from crashing and to have all the tiles and tile
behaviours updated in real time.

+++ Assets/Scripts/EnemySpawner.cs
@@ -6,6 +6,7 @@ public class EnemySpawner : MonoBehaviour
 
 	public GameObject enemy;                // The enemy prefab to be spawned.
 	public float spawnTime = 12f;            // How long between each spawn.
+	public int enemyCount; 
 	public Transform[] spawnPoints;         // An array of the spawn points this enemy can spawn from.
 	public List<Transform> tiles;
 	bool spawn;
@@ -14,6 +15,7 @@ public class EnemySpawner : MonoBehaviour
 	{
 		// Call the Spawn function after a delay of the spawnTime and then continue to call after the same amount of time.
 	//	InvokeRepeating ("Spawn", spawnTime, spawnTime);
+		enemyCount = 0;
 		spawn = true;
 	}
 
@@ -41,32 +43,33 @@ public class EnemySpawner : MonoBehaviour
 	
 	void Spawn ()
 	{
+		if (enemyCount <= 10) {
+			tiles = new List<Transform> ();
 
-		tiles = new List<Transform>();
-
-		//ArrayList tiles = new ArrayList ();
-		Transform transform = GameObject.Find ("HexagonGrid").transform;
-		foreach (Transform child in transform) {
-			//if(child.gameObject.GetComponent<TileBehaviour>().isPassable == true)
-			tiles.Add(child);
-		}
-		tiles.RemoveRange (22, (tiles.Count - 22));
-		tiles.ToArray ();
+			//ArrayList tiles = new ArrayList ();
+			Transform transform = GameObject.Find ("HexagonGrid").transform;
+			foreach (Transform child in transform) {
+				//if(child.gameObject.GetComponent<TileBehaviour>().isPassable == true)
+				tiles.Add (child);
+			}
+			tiles.RemoveRange (22, (tiles.Count - 22));
+			tiles.ToArray ();
 		
-		int randomTileIndex = UnityEngine.Random.Range (0, 22);
-		GameObject spawnTile = tiles [randomTileIndex].gameObject;
-		Vector3 position = new Vector3 (spawnTile.transform.position.x, 0.12f, spawnTile.transform.position.z);
-		// Create an instance of the enemy prefab at the randomly selected spawn point's position and rotation.
+			int randomTileIndex = UnityEngine.Random.Range (0, 22);
+			GameObject spawnTile = tiles [randomTileIndex].gameObject;
+			Vector3 position = new Vector3 (spawnTile.transform.position.x, 0.12f, spawnTile.transform.position.z);
+			// Create an instance of the enemy prefab at the randomly selected spawn point's position and rotation.
 
-		Transform myUnitTransform = enemy.transform.FindChild("Enemy");
+			Transform myUnitTransform = enemy.transform.FindChild ("Enemy");
 		
-		spawnTile.gameObject.GetComponent<TileBehaviour>().isEnemy = true;
-		spawnTile.gameObject.GetComponent<TileBehaviour>().isPassable = false;
-		myUnitTransform.GetComponent<CharacterMovement>().unitOriginalTile = spawnTile.gameObject.GetComponent<TileBehaviour> ();
-		Instantiate (enemy, position, Quaternion.identity);
-		SpawnDone ();
+			spawnTile.gameObject.GetComponent<TileBehaviour> ().isEnemy = true;
+			spawnTile.gameObject.GetComponent<TileBehaviour> ().isPassable = false;
+			myUnitTransform.GetComponent<CharacterMovement> ().unitOriginalTile = spawnTile.gameObject.GetComponent<TileBehaviour> ();
+			Instantiate (enemy, position, Quaternion.identity);
+			enemyCount += 1;
+			SpawnDone ();
 
-		/*
+			/*
 		position.y += 0.11f;
 		RaycastHit objectHit;
 		Vector3 down = Vector3.down;
@@ -98,6 +101,7 @@ public class EnemySpawner : MonoBehaviour
 			}
 		}
 */
-	//	enemy.GetComponent<CharacterMovement>().unitOriginalTile;
+			//	enemy.GetComponent<CharacterMovement>().unitOriginalTile;
+		}
 	}
 }
